feat: further registrar refactor
